Integrate DynamicRLEList as freeSpaceIndex

+++ src/main/java/de/uni_koblenz/west/koral/master/statisticsDB/impl/multi_file/storage/ExtraRowStorage.java
@@ -8,7 +8,7 @@ public interface ExtraRowStorage extends RowStorage {
 
 	public void deleteRow(long rowId);
 
-	public long[] getFreeSpaceIndexData();
+	public byte[] getFreeSpaceIndexData();
 
 	public void defragFreeSpaceIndex();
